+++ tests/run-night-testset.js
@@ -62,7 +62,7 @@ function runNightTestset()
       params.debugFrameOverlays = true;
       params.matchDiagnostics = true;
       params.nightShowPredicted = true;
-      params.tleExtraGroups = [ "classfd", "last-30-days", "cosmos-2251-debris",
+      params.tleExtraGroups = [ "classfd", "last-30-days", "catalog", "cosmos-2251-debris",
                                 "iridium-33-debris", "fengyun-1c-debris",
                                 "cosmos-1408-debris" ];